refactor(update): address review — drop re-export indirection, narrow excepts
- Repoint consumers (shell __init__, /update, /settings) and tests to import
  auto_update_enabled / auto_update_override_reason directly from the canonical
  `update_policy` module, and drop the unused re-export shims from
  ui/shell/update.py (resolves "unused import" findings).
- Narrow the broad `except Exception` in `info._auto_update_info` to
  (OSError, ValueError, ImportError) and log the degraded path instead of
  silently swallowing (C03); ConfigError/pydantic errors are ValueError.
- Narrow `update_policy.is_running_from_source_checkout` to
  (ImportError, AttributeError, OSError).
- Strengthen the /update auto persist tests to assert real on-disk persistence
  via load_config rather than mock call-coupling.
